backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
qdrant-client==1.8.0
google-generativeai==0.4.0
litellm==1.35.12
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2